Changed the window size initialization flow & Fixed the home button crash bug
Instead of initializing the window at a fixed size, try obtaining the screen size and setting it to 70% of the screen height (50% on Mac because of Retina's 2x display), and 70% of screen * 1.618 as the width.

Fixed the Home button crash bug, now it has a "plotted" flag in our PlotCanvas class

Changed the slider connect to valueChanged, this makes the slider and the date edit only allow dates in intervals, does not allow specific input in dates.